+++ src/main/java/org/spongepowered/common/mixin/core/world/extent/MixinExtent.java
@@ -163,22 +163,22 @@ public abstract class MixinExtent implements Extent {
     }
 
     @Override
-    public byte getLuminance(Vector3i position) {
+    public int getLuminance(Vector3i position) {
         return getLuminance(position.getX(), position.getY(), position.getZ());
     }
 
     @Override
-    public byte getLuminance(int x, int y, int z) {
-        return (byte) Math.max(getLuminanceFromGround(x, y, z), getLuminanceFromSky(x, y, z));
+    public int getLuminance(int x, int y, int z) {
+        return Math.max(getLuminanceFromGround(x, y, z), getLuminanceFromSky(x, y, z));
     }
 
     @Override
-    public byte getLuminanceFromSky(Vector3i position) {
+    public int getLuminanceFromSky(Vector3i position) {
         return getLuminanceFromSky(position.getX(), position.getY(), position.getZ());
     }
 
     @Override
-    public byte getLuminanceFromGround(Vector3i position) {
+    public int getLuminanceFromGround(Vector3i position) {
         return getLuminanceFromGround(position.getX(), position.getY(), position.getZ());
     }
